requirements.txt
# Web Framework
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6

# Image Processing
pillow==9.5.0
numpy

# PyTorch
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.0+cpu
torchvision==0.17.0+cpu